Create a ::Torrent::read, support for specifying fields
This change adds the possibility to specify only a limited set of fields, to be
able to take advantage of only sending one RPC request, while still not having
to request everything. This change introduces the ->read method to the
Transmission::Client class. The existing ->read_all method simply wraps the new
read method.

This change also introduces a dependency on List::MoreUtils; this module should
already be required by Moose.