types: fix perl precedence warning in _is_num
This change simplifies the expression for determining what perl
internally thinks of the variable. It does this simplification on the
expectation that POK can't be true at the same time as IOK | NOK. The
unit tests pass, even for cases like "10.0", so it seems to be true at
least for perl 5.14.2 and 5.20.1.